chore: move pages into its own folder
