+++ src/view/pages/LandingPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { FaPlay } from "react-icons/fa";
-import { Button } from "../../../view/components/atomic/Button";
-import { PageTooltip } from "../../components/atomic/PageTooltip";
+import { Button } from "../components/atomic/Button";
+import { PageTooltip } from "../components/atomic/PageTooltip";
 
 export const LandingPage = ({
     onRecordButtonPress,
